Create backup.sh
Create backups for all of your EBS volumes and optionally delete all the previous ones. Backups are taken first so that the subsequent deletion of previous snapshots doesn't incur data loss. From the documentation, "When you delete a snapshot, only the data not needed for any other snapshot is removed. So regardless of which prior snapshots have been deleted, all active snapshots will have access to all the information needed to restore the volume." 

https://docs.aws.amazon.com/cli/latest/reference/ec2/delete-snapshot.html